embedding-service/requirements.txt
flask==2.3.3
transformers==4.36.2
torch==2.1.2
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.3.1
python-dotenv==1.0.0
flask-cors==4.0.0 